CSPACE-5359: findDocs now accepts an ORDER BY statement, while still defaulting to returning documents in descending order by updatedAt timestamps. When updating display names in authority references, we now obtain documents in ascending order by createdAt timestamps, which should be static for existing records.
Conflicts:

	services/common/src/main/java/org/collectionspace/services/nuxeo/client/java/RepositoryJavaClientImpl.java
	services/common/src/main/java/org/collectionspace/services/nuxeo/util/NuxeoUtils.java

Some changes were made in the conflicting files to remove dependencies on other changes that were made post-2.4.

A change was also made to:

services/servicegroup/service/src/main/java/org/collectionspace/services/servicegroup/nuxeo/ServiceGroupDocumentModelHandler.java

This was to make the class compile, with the new RepositoryJavaClientImpl.findDocs signature required by CSPACE-5359.

When upgrading, all of the above changes can be discarded in favor of the 2.5+ versions of the files.