Aggregate repeating fields
This replaces many of the simple joins with subqueries which run string_agg to combine multiple rows of data. In addition, for dimension there are multiple levels of aggregation which can occur so we first aggregate grouping by the measured part id, then join for each parent id. Likely not very performant.